fars_map_state {TestPackage}	R Documentation

Fars Map States

Description

This functions takes a year and state number as inputs and returns a plot of the fatalities plotted on a graphic of the state.

Usage


fars_map_state(state.num, year)

Arguments

state.num
a state number from 1 to 50 representing one of the states of the United States.

year
a year in the format of "yyyy" (multiple years cannot be supplied to this funtion)

Value

This function returns a plot of the state indicated with all fatalities plotted by latitude/longitude. An error will be return if an invalid state number is supplied, or the year supplied is not found in the

Examples


## Not run: 
fars_map_state(25,2013)

## End(Not run)


[Package TestPackage version 0.0.0.9000 Index]