fars_read {TestPackage}	R Documentation

Fars Read

Description

This is a simple function that reads in a CSV file and returns a tbl_df/dataframe as it's output.

Usage


fars_read(filename)

Arguments

filename
A character string of the name of the CSV file that you'd like to read into a dataframe

Value

This function returns the specified file CSV file as a tbl_df/dataframe object. An error will be returned if the filename is unable to be found in the working directory.

Examples


## Not run: 
fars_read("example.csv")

## End(Not run)

[Package TestPackage version 0.0.0.9000 Index]