fars_read_years {TestPackage}	R Documentation

Fars Read Years

Description

This functions takes years as inputs and returns a list containing the year/month combinations found in the FAR files

Usage


fars_read_years(years)

Details

The functions uses the Fars_Read/Make_Filename functions to take the year provided and create a list of the months/year combinations if found in the FAR files.

Value

This function returns a list of the year/month found in the FAR file based on the years provided. It will return an error if the year provided cannot be converted into an integer format or if the filename resulting from the years specified cannot be found in the directory.

Examples


## Not run: 
fars_read_years(2016)
fars_read_years(c(2014,2013))

## End(Not run)

[Package TestPackage version 0.0.0.9000 Index]