fars_summarize_years {TestPackage}	R Documentation

Fars Summarize Years

Description

This functions takes years as inputs and returns a tbl_df/dataframe object totaling the number of fatalities per month for each year specified.

Usage


fars_summarize_years(years)

Details

The functions uses the fars_read, make_filename, fars_read_years functions. This function requires the following external package functions:

Value

This function returns a tbl_dataframe object totaling the number of fatalities per month for each year specified. An error will be returned if a year provided cannot be converted to integer format, or if no file corresponding to the year specified is found in the working directory.

Examples


## Not run: 
fars_summarize_years(2016)
fars_summarize_years(c(2014,2013))

## End(Not run)

[Package TestPackage version 0.0.0.9000 Index]