make_filename {TestPackage}	R Documentation

Make Filename

Description

This function converts a year into the properly formatted filename

Usage


make_filename(year)

Arguments

year
a year in the format of "yyyy"

Value

This function returns the specified filename as a string to be used with the entered year. This function returns an error if value entered cannot be converted properly into an integer format.

Examples


## Not run: 
make_filename(2016)

## End(Not run)

[Package TestPackage version 0.0.0.9000 Index]